// apps/web-admin/src/interfaces/adminInterface.ts
enum AdminType {
  SUPER = "SUPER",
  NORMAL = "NORMAL",
}

export interface IAdmin {
  admin_id: string;
  name: string;
  profile_picture: string;
  email: string;
  username: string;
  password: string;
  contact_number: string;
  address: string;
  date_of_birth: Date;
  admin_type: AdminType;
}